Reoptimized Inventory search algorithm
Instead of 2 foreach each loops constantly checking each and every element in the last and comparing it to the one that is about to be added and constanly checking is the element is stackable. I have optimized the code to use only one foreach loopto check for the elements and keep track of the first available slot it finds to store the element incase no stacking is involved. This has greatly reduced the complexity of the algorithm and has made it more performant on low cpu devices.